backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
python-dotenv>=1.0.0
pydantic[email]>=2.10.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
